requirements.txt
pygame-ce
pytest
black
ruff
mypy
